requirements.txt
flask
flask-restx
psycopg2-binary